Refactor ProcessCubeLibrary and extract Keywords into base-classes
