+++ requirements.txt
@@ -1,4 +1,4 @@
-atlas-engine-client>=2.0.0a4
+atlas-engine-client>=2.0.0a7
 docker==5.0.0
 requests==2.24.0
 robotframework>=4.0.0
